--- src/lex.rs
@@ -1,4 +1,0 @@
-  
-pub fn mylex() -> String {
-  "lexed1".to_string()
-}
